+++ .prettierrc.cjs
@@ -2,7 +2,7 @@ const config = require('@fundamend/config-prettier');
 
 module.exports = {
 	...config,
-	plugins: ["prettier-plugin-svelte"],
-	pluginSearchDirs: ["."],
-	overrides: [{ "files": "*.svelte", "options": { "parser": "svelte" } }]
+	plugins: ['prettier-plugin-svelte'],
+	pluginSearchDirs: ['.'],
+	overrides: [{ files: '*.svelte', options: { parser: 'svelte' } }]
 };