+++ .stylelintrc.cjs
@@ -3,6 +3,6 @@ const fundamend = require('@fundamend/config-stylelint');
 module.exports = {
 	...fundamend,
 	rules: {
-		"no-descending-specificity": null
+		'no-descending-specificity': null
 	}
 };