+++ functions/api/v1/private/_middleware.js
@@ -1,3 +1,0 @@
-import verifyJWTPlugin from "@fundamend/pages-plugin-verify-jwt";
-
-export const onRequest = verifyJWTPlugin();