+++ packages/app/src/routes/(private)/account/+page.svelte
@@ -1,0 +1,10 @@
+<script>
+	import { onMount } from 'svelte';
+
+	onMount(async () => {
+		const auth = await (await import('$lib/auth.js')).default;
+		await auth.userProfile();
+	});
+</script>
+
+<p>redirecting...</p>