+++ packages/app/src/routes/(private)/private/+page.svelte
@@ -1,0 +1,5 @@
+<fundamend-box class="width:100 height:100">
+	<fundamend-center>
+		<h1>Private</h1>
+	</fundamend-center>
+</fundamend-box>