+++ packages/app/src/routes/(public)/+layout.svelte
@@ -1,0 +1,9 @@
+<div class="indicator" />
+
+<slot />
+
+<style>
+	.indicator {
+		border: 1px solid green;
+	}
+</style>