+++ src/routes/(public)/+layout.svelte
@@ -1,9 +1,0 @@
-<div class="indicator"></div>
-
-<slot></slot>
-
-<style>
-	.indicator {
-		border: 1px solid green;
-	}
-</style>