+++ src/routes/(public)/public/+page.svelte
@@ -1,6 +1,0 @@
-<fundamend-box class="width:100 height:100">
-	<fundamend-center>
-		<h1>Public</h1>
-	</fundamend-center>
-</fundamend-box>
-