+++ CHANGELOG.md
@@ -4,8 +4,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### 0.0.1 (2022-10-22)
 
-
 ### Features
 
-* add test page function ([eb25e06](https://github.com/fundamend/fundamend/commit/eb25e06f136d8ae1ce5b93dac71be42589449a88))
-* scaffold sveltekit ([149405d](https://github.com/fundamend/fundamend/commit/149405d29831928d6b24b1a507ad9bd09cc2895e))
+- add test page function ([eb25e06](https://github.com/fundamend/fundamend/commit/eb25e06f136d8ae1ce5b93dac71be42589449a88))
+- scaffold sveltekit ([149405d](https://github.com/fundamend/fundamend/commit/149405d29831928d6b24b1a507ad9bd09cc2895e))